All admin metods with validation

+++ src/main/java/com/example/api/service/AdminService.java
@@ -1,16 +1,12 @@
-
 package com.example.api.service;
 
-import com.example.api.entity.Description;
 import com.example.api.entity.Product;
-import com.example.api.model.ProductModel;
+import com.example.api.exeptons.IdNotFoundException;
 import com.example.api.repository.DescriptionRepository;
 import com.example.api.repository.PriceRepository;
 import com.example.api.repository.ProductRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
-import java.util.List;
 
 @Service
 public class AdminService {
@@ -22,31 +18,40 @@ public class AdminService {
     @Autowired
     PriceRepository priceRepository;
 
-    public Product addProduct(ProductModel productModel){
-        Product product = new Product();
-        productModel.getDescriptions().forEach((d) -> product.addDescription(d));
-        productModel.getPrices().forEach((p) -> product.addPrice(p));
-        return productRepository.save(product);
+    public Product addProduct(Product product){
+        Product productNew = new Product();
+        product.getDescriptions().forEach((d) -> productNew.addDescription(d));
+        product.getPrices().forEach((p) -> productNew.addPrice(p));
+        return productRepository.save(productNew);
     }
 
     public Iterable<Product> getAll(){
         return productRepository.findAll();
     }
 
-    public Product getById(Long id){
+    public Product getById(Long id) throws IdNotFoundException {
+        if (!productRepository.existsById(id)) {
+            throw new IdNotFoundException("");
+        }
         return productRepository.findById(id).get();
     }
 
-    public void updateProduct(Long id , ProductModel productModel){
-        Product product = productRepository.findById(id).get();
+    public void updateProduct(Long id , Product product) throws IdNotFoundException {
+        if (!productRepository.existsById(id)) {
+            throw new IdNotFoundException("");
+        }
+        Product productNew = productRepository.findById(id).get();
         product.getDescriptions().clear();
         product.getPrices().clear();
-        productModel.getDescriptions().forEach((d) -> product.addDescription(d));
-        productModel.getPrices().forEach((p) -> product.addPrice(p));
-        productRepository.save(product);
+        product.getDescriptions().forEach((d) -> productNew.addDescription(d));
+        product.getPrices().forEach((p) -> productNew.addPrice(p));
+        productRepository.save(productNew);
     }
 
-    public void deleteProduct(Long id){
+    public void deleteProduct(Long id) throws IdNotFoundException {
+        if (!productRepository.existsById(id)) {
+            throw new IdNotFoundException("");
+        }
         productRepository.deleteById(id);
     }
 }
